requirements.txt
torch>=2.2.0
torchvision>=0.17.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
streamlit>=1.33.0
Pillow>=10.2.0
PyYAML>=6.0.1
joblib>=1.3.0
pytest>=8.1.0
